Introduce "trustedRedirects" flag
This commit adds the possibility to specify a comma-separated string of trusted hosts and ip addresses with the "--trustedRedirects" flag.
Normally, request are automatically redirected to hosts but without the authorization header. With this fix, redirects from non-leading
masters to the leading master will now set the authorization header when it is required. Additionally, redirects to non-trusted locations
 will be prohibited by default.